back/requirements.txt
flask-swagger-ui==4.11.1
mysql-connector-python==8.2.0
pip-chill==1.0.3
py-bcrypt==0.4
pyjwt==2.8.0
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
